Split logic into encoder/decoder units and optimize core loop
However, this model is still insufficient, because it employs:

  1-input, n-handlers, n-outputs

And the bottlenecks appear around applying n-handlers to 1 input
( particularly, lots of time is lost in cloning every record from input
  for each handler, so that handlers modifying the state to relay to
  output doesn't result in sibling handlers getting corrupted )

What we need is a system where we can do:

  1-input, 1-handler, n-outputs

where unnnecessary handler calls and matches can be elided.

Unfortunately, the logic to pull that off is .... not easy.